secondsChange and secondsStay type is Double

+++ app/src/main/java/com/example/dipolia/presentation/StreamingState.kt
@@ -3,6 +3,6 @@ package com.example.dipolia.presentation
 data class StreamingState(
 
     var isLooping: Boolean? = null,
-    var secondsChange: Int? = null,
-    var secondsStay: Int? = null
+    var secondsChange: Double? = null,
+    var secondsStay: Double? = null
 )
